pyproject.toml
[project]
name = "eda-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browser-use>=0.5.9",
    "e2b-code-interpreter>=1.5.2",
    "matplotlib>=3.10.5",
    "openai>=1.98.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
